src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.14.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.2.0
    # via -r requirements.in
black==24.8.0
    # via -r requirements.in
boto3==1.34.154
    # via
    #   -r requirements.in
    #   moto
boto3-stubs==1.34.156
    # via types-boto3
botocore==1.34.154
    # via
    #   boto3
    #   moto
    #   pynamodb
    #   s3transfer
botocore-stubs==1.34.156
    # via boto3-stubs
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   typer
    #   uvicorn
coverage[toml]==7.6.1
    # via pytest-cov
cryptography==43.0.0
    # via moto
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
execnet==2.1.1
    # via pytest-xdist
fastapi[all]==0.112.0
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
flake8==7.1.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   pytest-httpx
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isort==6.0.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via
    #   fastapi
    #   moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mako==1.3.8
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
    #   werkzeug
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
moto==5.0.12
    # via -r requirements.in
mypy==1.15.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
orjson==3.10.6
    # via fastapi
packaging==24.1
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements.in
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.20.1
    # via pydantic
pydantic-extra-types==2.9.0
    # via fastapi
pydantic-settings==2.4.0
    # via fastapi
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pyjwt==2.9.0
    # via -r requirements.in
pynamodb==6.0.1
    # via -r requirements.in
pytest==8.3.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-httpx
    #   pytest-xdist
pytest-asyncio==0.23.8
    # via -r requirements.in
pytest-cov==5.0.0
    # via -r requirements.in
pytest-httpx==0.30.0
    # via -r requirements.in
pytest-xdist==3.6.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-json-logger==2.0.7
    # via -r requirements.in
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   fastapi
    #   responses
    #   uvicorn
requests==2.32.3
    # via
    #   -r requirements.in
    #   moto
    #   responses
responses==0.25.3
    # via moto
rich==13.7.1
    # via typer
s3transfer==0.10.2
    # via boto3
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.37
    # via
    #   -r requirements.in
    #   alembic
    #   sqlalchemy-utils
    #   sqlmodel
sqlalchemy-utils==0.41.2
    # via -r requirements.in
sqlmodel==0.0.22
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
types-awscrt==0.21.2
    # via botocore-stubs
types-boto3==1.0.2
    # via -r requirements.in
types-pyyaml==6.0.12.20240808
    # via -r requirements.in
types-requests==2.32.0.20240712
    # via -r requirements.in
types-s3transfer==0.10.1
    # via boto3-stubs
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
ujson==5.10.0
    # via fastapi
urllib3==2.2.2
    # via
    #   botocore
    #   requests
    #   responses
    #   types-requests
uvicorn[standard]==0.30.5
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
werkzeug==3.0.3
    # via moto
xmltodict==0.13.0
    # via moto
